refactor: split executor and reporter output
